:boom: Combine Validation and UX script generation
Resolves https://github.com/punchcard-cms/content-types/issues/43
