+++ lib/form.js
@@ -1,12 +1,8 @@
 'use strict';
 
 const html = require('./form/html');
-const validation = require('./form/validation');
 const scripts = require('./form/scripts');
 
-const validate = require('./form/validate');
-
-// const scripts = require('./form/scripts');
 // const css = require('./form/css');
 
 
@@ -15,10 +11,6 @@ const form = (type, errors) => {
 
   return html(type, errors).then(result => {
     rendered.html = result;
-
-    return validation(type);
-  }).then(result => {
-    rendered.validation = result;
 
     return scripts(type);
   }).then(result => {
@@ -30,4 +22,4 @@ const form = (type, errors) => {
 
 
 module.exports = form;
-module.exports.validate = validate;
+module.exports.scripts = scripts;
