+++ lib/form/js/validation.js
@@ -1,0 +1,91 @@
+/* eslint-env browser */
+/* global allPlugins, allInputs, allSettings, allIDs */
+
+(function formValidation() {
+  var valueCheck = function (input, settings, validation) {
+    if (input.target.required && (input.target.required === 'save' || input.target.required === 'publish') && input.target.value === '') {
+      return 'Field cannot be left blank!';
+    }
+
+    return validation(input, settings);
+  };
+
+  // Validation Function
+  var validate = function (e) {
+    var target = e.target;
+    var value = target.value;
+    var inp = allInputs[target.id];
+    var parent = allIDs[inp.parent.id];
+    var input = {};
+    var settings = {};
+    var validation = false;
+    var message;
+    var currentMessage = parent.querySelector('[role="alert"][for="' + target.id + '"]');
+    var insertRef;
+
+    // Set up validation input
+    input.target = {
+      value: value,
+      name: inp.name,
+      required: inp.required,
+    };
+    input.all = {};
+    input.all[inp.name] = value;
+
+    // Set up validation settings
+    settings.target = inp.settings;
+    settings.all = {};
+    settings.all[inp.name] = inp.settings;
+
+    // Set all sibling input and settings
+    Object.keys(inp.siblings).forEach(function (sibling) {
+      input.all[sibling] = allIDs[inp.siblings[sibling]].value;
+      settings.all[sibling] = allSettings[inp.parent.id][sibling];
+    });
+
+    // Run the validation
+    validation = valueCheck(input, settings, allPlugins[inp.parent.type].validation[inp.validation.function]);
+
+    // Add/Remove validation
+    if (validation === true) {
+      // Check to see if there is a current alert message for this input
+      if (currentMessage) {
+        // Remove invalid
+        target.removeAttribute('aria-invalid');
+
+        // Delete the current message if it exists
+        parent.removeChild(currentMessage);
+      }
+    }
+    else {
+      // Create error message
+      message = document.createElement('p');
+      message.className = 'form--alert';
+      message.setAttribute('role', 'alert');
+      message.setAttribute('for', target.id);
+      message.textContent = validation;
+
+      // Check to see if there is a current alert message for this input
+      if (!message.isEqualNode(currentMessage)) {
+        // Set element to invalid
+        target.setAttribute('aria-invalid', 'true');
+
+        if (parent.nodeName.toLowerCase() !== 'fieldset') {
+          // If parent isn't a fieldset, add message before first child
+          insertRef = parent.firstChild;
+          parent.insertBefore(message, insertRef);
+        }
+        else {
+          // If parent is a fieldset, add message after either the description, if it exists, or the legend
+          insertRef = parent.querySelector('.form--description') || parent.querySelector('legend');
+          parent.insertBefore(message, insertRef.nextSibling);
+        }
+      }
+    }
+  };
+
+  // Add all event listeners
+  Object.keys(allInputs).forEach(function (id) {
+    allIDs[id].addEventListener(allInputs[id].validation.on, validate);
+  });
+}());
