+++ lib/form/scripts.js
@@ -2,14 +2,22 @@
 
 const browserify = require('browserify');
 const stream = require('stream');
+const fs = require('fs');
+const path = require('path');
+
+const formJS = fs.readdirSync(path.join(__dirname, 'js')).map(script => {
+  return fs.readFileSync(path.join(__dirname, 'js', script));
+}).join('\n\n');
+
+// const util = require('../util');
 
 const rendered = (type) => {
   return new Promise((res, rej) => {
     const b = browserify();
     const s = new stream.Readable();
 
-    const scriptFunctions = {};
-    const allScriptSettings = {};
+    const allPlugins = {};
+    const allSettings = {};
     const inputs = {};
     const ids = {};
     let plugins = type.attributes.map(plugin => {
@@ -23,14 +31,14 @@ const rendered = (type) => {
     plugins.forEach(plugin => {
       // Require Plugins for Browserify
       b.require(`input-plugin-${plugin}`);
-      scriptFunctions[plugin] = `input-plugin-${plugin}`;
+      allPlugins[plugin] = `input-plugin-${plugin}`;
     });
 
     type.attributes.map(plugin => {
       const children = Object.keys(plugin.inputs);
 
       // All settings for this plugin
-      allScriptSettings[plugin.id] = {};
+      allSettings[plugin.id] = {};
 
       // Add ID to all IDs
       ids[plugin.id] = plugin.id;
@@ -49,23 +57,25 @@ const rendered = (type) => {
         });
 
         // Set settings for this input blank as default
-        allScriptSettings[plugin.id][input] = {};
+        allSettings[plugin.id][input] = {};
 
         // Set to actual settings if it exists
         if (inp.hasOwnProperty('settings')) {
-          allScriptSettings[plugin.id][input] = inp.settings;
+          allSettings[plugin.id][input] = inp.settings;
         }
 
         // Build input object for this specific input
         inputs[inp.id] = {};
         inputs[inp.id].name = input;
+        inputs[inp.id].validation = inp.validation;
         inputs[inp.id].script = inp.script;
+        inputs[inp.id].required = inp.required;
         inputs[inp.id].settings = {};
-        inputs[inp.id].siblings = {};
         inputs[inp.id].parent = {
           id: plugin.id,
           type: plugin.type,
         };
+        inputs[inp.id].siblings = {};
 
         if (inp.hasOwnProperty('settings')) {
           inputs[inp.id].settings = inp.settings;
@@ -81,65 +91,35 @@ const rendered = (type) => {
       });
     });
 
-    const scripts = `
-var allScripts = ${JSON.stringify(scriptFunctions)};
-var allScriptInputs = ${JSON.stringify(inputs)};
-var allScriptSettings = ${JSON.stringify(allScriptSettings)};
-var allScriptIDs = ${JSON.stringify(ids)};
+    const formScript = `
+var allPlugins = ${JSON.stringify(allPlugins)};
+var allInputs = ${JSON.stringify(inputs)};
+var allSettings = ${JSON.stringify(allSettings)};
+var allIDs = ${JSON.stringify(ids)};
 
 // Work those requires
-Object.keys(allScripts).forEach(function(plugin) {
-  var plug = require(allScripts[plugin]);
+Object.keys(allPlugins).forEach(function(plugin) {
+  var plug = require(allPlugins[plugin]);
 
-  // Require the script from the inputs
-  allScripts[plugin] = plug.scripts;
+  // Require the plugin
+  allPlugins[plugin] = plug;
 });
 
-// Script Function
-var script = function (e) {
-  var target = e.target;
-  var inp = allScriptInputs[target.id];
-  var input = {};
-  var settings = {};
-
-  // Set up script input
-  input.target = target;
-  input.all = {};
-  input.all[inp.name] = target;
-
-  // Set up script settings
-  settings.target = inp.settings;
-  settings.all = {};
-  settings.all[inp.name] = inp.settings;
-
-  // Set all sibling input and settings
-  Object.keys(inp.siblings).forEach(function (sibling) {
-    input.all[sibling] = allScriptIDs[inp.siblings[sibling]];
-    settings.all[sibling] = allScriptSettings[inp.parent.id][sibling];
-  });
-
-  // Run the script
-  script = allScripts[inp.parent.type][inp.script.function](input, settings);
-
-};
 
 // Add Document Event Listener
 document.addEventListener('DOMContentLoaded', function () {
   // Get all elements so we don't need to constantly read from the DOM
-  Object.keys(allScriptIDs).forEach(function (id) {
-    allScriptIDs[id] = document.getElementById(id);
+  Object.keys(allIDs).forEach(function (id) {
+    allIDs[id] = document.getElementById(id);
   });
 
-  // Add all event listeners
-  Object.keys(allScriptInputs).forEach(function (id) {
-    if (allScriptInputs[id].hasOwnProperty('script')) {
-      allScriptIDs[id].addEventListener(allScriptInputs[id].script.on, script);
-    }
-  });
+  // Inlcude all form scripts
+  ${formJS}
 });`;
+
     s._read = function read() {};
 
-    s.push(scripts);
+    s.push(formScript);
     s.push(null);
 
     b.add(s);
